feat: [CDS-0]: move code from resource to service impl
